% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrinsic_incubation_period.R
\name{extrinsic_incubation_period}
\alias{extrinsic_incubation_period}
\title{Extrinsic incubation period}
\usage{
extrinsic_incubation_period(temp, temp_min = 16)
}
\arguments{
\item{temp}{\code{numeric} value of temperature for EIP to be calculated}

\item{temp_min}{\code{numeric} minimum temperature for incubation}
}
\value{
\code{numeric} calculation of EIP
}
\description{
Extrinsic incubation period
}
\details{
Extrinsic Incubation Period function depending on T. Below T_min the parasite cannot complete sporogony; return \code{Inf} so downstream p^n collapses to 0 (VC -> 0) instead of going negative.
}
\examples{

temps <- seq(0, 45, by = 1)

extrinsic_incubation_period(temp = temps, temp_min = 14)

}
